feat: replace (platform) deploy + start instance modals with overlays
Related to #2552
