--- client/src/plugins/camunda-plugin/deployment-tool/DeploymentConfigModal.less
@@ -1,18 +1,0 @@
-:local(.DeploymentConfigModal) {
-  .icon {
-    vertical-align: text-top;
-  }
-
-  .intro {
-    margin-bottom: 30px;
-  }
-
-  fieldset:first-child {
-    margin-top: 0px;
-  }
-
-  fieldset:last-child {
-    margin-bottom: 0px;
-    padding-bottom: 16px;
-  }
-}
