+++ client/src/plugins/camunda-plugin/deployment-tool/DeploymentConfigOverlay.less
@@ -1,0 +1,19 @@
+:local(.DeploymentConfigOverlay) {
+  .icon {
+    vertical-align: text-top;
+  }
+
+  .intro {
+    margin: 0 0 16px 0;
+    font-size: 13px;
+    color: var(--color-grey-225-10-55);
+  }
+
+  .toggle-switch {
+    margin-top: 14px;
+  }
+
+  .toggle-switch__label {
+    font-size: 13px !important;
+  }
+}
