--- client/src/plugins/camunda-plugin/start-instance-tool/StartInstanceConfigModal.less
@@ -1,5 +1,0 @@
-:local(.StartInstanceDetailsModal) {
-  .intro {
-    margin-bottom: 20px;
-  }
-}
