+++ client/src/plugins/camunda-plugin/start-instance-tool/StartInstanceConfigOverlay.less
@@ -1,0 +1,9 @@
+:local(.StartInstanceDetailsOverlay) {
+
+  .intro {
+    margin: 0 0 4px 0;
+    font-size: 13px;
+    color: var(--color-grey-225-10-55);
+  }
+    
+}
