+++ client/src/shared/ui/overlay/Overlay.less
@@ -65,12 +65,12 @@
     border-top: 1px solid var(--overlay-border-color);
   }
   
-  .fieldset {
+  fieldset {
     border: none;
-    margin: 0;
+    margin: 0 0 5px 0;
     padding: 0;
   }
-
+  
   .fields {
     display: grid;
     grid-gap: 12px;
